requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.3
django-allauth==0.50.0
django-braces==1.15.0
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-embed-video==1.4.4
django-environ==0.4.5
django-oauth-toolkit==2.0.0
django-picklefield==3.0.1
django-rest-framework-social-oauth2==1.1.0
django-sms==0.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-social-oauth2==1.2.1
ecdsa==0.17.0
facebook-sdk==3.1.0
future==0.18.2
google==3.0.0
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.2
httplib2==0.20.4
idna==3.3
jsonfield==3.1.0
jwcrypto==1.3.1
oauthlib==3.2.0
Pillow==9.1.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.9.9
PyJWT==2.3.0
pyotp==2.6.0
pyparsing==3.0.9
python-jose==3.3.0
python-twitter==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
twilio==7.9.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
